refactor(auth): extract shared leaf helpers
