+++ src/auth/shared/safari-auth-fallback.js
@@ -1,0 +1,19 @@
+// iOS standalone PWA Safari fallback: armed after a failed attempt,
+// then the next Login tap opens the app URL in Safari from a user gesture.
+let safariExternalOpenArmed = false;
+
+/**
+ * Get whether Safari external open fallback is armed.
+ * @returns {boolean}
+ */
+export function isSafariExternalOpenArmed() {
+  return safariExternalOpenArmed;
+}
+
+/**
+ * Set Safari external open fallback armed state.
+ * @param {boolean} value
+ */
+export function setSafariExternalOpenArmed(value) {
+  safariExternalOpenArmed = value;
+}
